[refactor] Usecase: categorise usecases
