+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/usecase/account/InitialiseAccountUseCase.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.account
 
 import com.rwmobi.kunigami.domain.exceptions.NoValidMeterException
 import com.rwmobi.kunigami.domain.exceptions.except
