+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/usecase/account/UpdateMeterPreferenceUseCase.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.account
 
 import co.touchlab.kermit.Logger
 import com.rwmobi.kunigami.domain.exceptions.except
