+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/usecase/product/GetLatestProductByKeywordUseCase.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.product
 
 import co.touchlab.kermit.Logger
 import com.rwmobi.kunigami.domain.repository.OctopusApiRepository
