+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/usecase/product/GetStandardUnitRateUseCase.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.product
 
 import com.rwmobi.kunigami.domain.exceptions.except
 import com.rwmobi.kunigami.domain.model.rate.Rate
