+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/usecase/product/GetTariffUseCase.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.product
 
 import com.rwmobi.kunigami.domain.model.product.Tariff
 import com.rwmobi.kunigami.domain.repository.OctopusApiRepository
