+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/account/SyncUserProfileUseCaseTest.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.account
 
 import com.rwmobi.kunigami.domain.exceptions.IncompleteCredentialsException
 import com.rwmobi.kunigami.domain.model.account.UserProfile
