+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/consumption/GetConsumptionAndCostUseCaseTest.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.consumption
 
 import com.rwmobi.kunigami.domain.model.consumption.Consumption
 import com.rwmobi.kunigami.domain.model.consumption.ConsumptionTimeFrame
