+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/product/GetFilteredProductsUseCaseTest.kt
@@ -1,4 +1,11 @@
-package com.rwmobi.kunigami.domain.usecase
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.kunigami.domain.usecase.product
 
 import com.rwmobi.kunigami.domain.model.product.ProductDirection
 import com.rwmobi.kunigami.domain.model.product.ProductFeature
@@ -11,29 +18,29 @@ import kotlinx.datetime.Instant
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertNull
+import kotlin.test.assertTrue
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class GetLatestProductByKeywordUseCaseTest {
+class GetFilteredProductsUseCaseTest {
 
-    private lateinit var getLatestProductByKeywordUseCase: GetLatestProductByKeywordUseCase
+    private lateinit var getFilteredProductsUseCase: GetFilteredProductsUseCase
     private lateinit var fakeRestApiRepository: FakeOctopusApiRepository
-    private val keyword = "AGILE"
+    private val samplePostcode = "WC1X 0ND"
 
     @BeforeTest
     fun setupUseCase() {
         fakeRestApiRepository = FakeOctopusApiRepository()
-        getLatestProductByKeywordUseCase = GetLatestProductByKeywordUseCase(
+        getFilteredProductsUseCase = GetFilteredProductsUseCase(
             octopusApiRepository = fakeRestApiRepository,
             dispatcher = UnconfinedTestDispatcher(),
         )
     }
 
     @Test
-    fun `GetLatestProductByKeywordUseCase should return tariff code successfully`() = runTest {
+    fun `invoke should return filtered products successfully`() = runTest {
         val productSummaries = listOf(
             ProductSummary(
-                code = "GO-VAR-22-10-14",
+                code = "1",
                 direction = ProductDirection.IMPORT,
                 fullName = "Product 1",
                 displayName = "Product 1",
@@ -44,7 +51,7 @@ class GetLatestProductByKeywordUseCaseTest {
                 brand = "OCTOPUS_ENERGY",
             ),
             ProductSummary(
-                code = "AGILE-24-04-03",
+                code = "2",
                 direction = ProductDirection.EXPORT,
                 fullName = "Product 2",
                 displayName = "Product 2",
@@ -55,7 +62,7 @@ class GetLatestProductByKeywordUseCaseTest {
                 brand = "OCTOPUS_ENERGY",
             ),
             ProductSummary(
-                code = "PREPAY-VAR-18-90-21",
+                code = "3",
                 direction = ProductDirection.IMPORT,
                 fullName = "Product 3",
                 displayName = "Product 3",
@@ -65,67 +72,42 @@ class GetLatestProductByKeywordUseCaseTest {
                 availability = Instant.parse("2023-01-01T00:00:00Z")..Instant.parse("2023-12-31T23:59:59Z"),
                 brand = "OCTOPUS_ENERGY",
             ),
-        )
-
-        fakeRestApiRepository.setProductsResponse = Result.success(productSummaries)
-
-        val result = getLatestProductByKeywordUseCase(keyword = keyword)
-
-        assertEquals(
-            expected = "AGILE-24-04-03",
-            actual = result,
-        )
-    }
-
-    @Test
-    fun `GetLatestProductByKeywordUseCase should return latest tariff code if more than one Agile products exist`() = runTest {
-        val productSummaries = listOf(
             ProductSummary(
-                code = "AGILE-FLEX-22-11-25",
-                direction = ProductDirection.EXPORT,
-                fullName = "Agile Octopus November 2022 v1",
-                displayName = "Agile Octopus",
-                description = "Sample description",
-                features = listOf(ProductFeature.VARIABLE),
+                code = "4",
+                direction = ProductDirection.IMPORT,
+                fullName = "Product 4",
+                displayName = "Product 4",
+                description = "Description 4",
+                features = listOf(ProductFeature.PREPAY),
                 term = 12,
-                availability = Instant.parse("2022-11-25T00:00:00Z")..Instant.parse("2023-12-11T12:00:00Z"),
-                brand = "OCTOPUS_ENERGY",
-            ),
-            ProductSummary(
-                code = "AGILE-24-04-03",
-                direction = ProductDirection.EXPORT,
-                fullName = "Agile Octopus April 2024 v1",
-                displayName = "Agile Octopus",
-                description = "Sample description",
-                features = listOf(ProductFeature.VARIABLE),
-                term = 12,
-                availability = Instant.parse("2024-04-03T00:00:00+01:00")..Instant.DISTANT_FUTURE,
-                brand = "OCTOPUS_ENERGY",
+                availability = Instant.parse("2023-01-01T00:00:00Z")..Instant.parse("2023-12-31T23:59:59Z"),
+                brand = "OTHER_BRAND",
             ),
         )
 
         fakeRestApiRepository.setProductsResponse = Result.success(productSummaries)
 
-        val result = getLatestProductByKeywordUseCase(keyword = keyword)
+        val result = getFilteredProductsUseCase(postcode = samplePostcode)
 
-        assertEquals(
-            expected = "AGILE-24-04-03",
-            actual = result,
-        )
+        assertTrue(result.isSuccess)
+        assertEquals(listOf(productSummaries[0]), result.getOrNull())
     }
 
     @Test
-    fun `GetLatestProductByKeywordUseCase should return null when repository call fails`() = runTest {
+    fun `invoke should return failure result when repository call fails`() = runTest {
         val errorMessage = "API Error"
         fakeRestApiRepository.setProductsResponse = Result.failure(RuntimeException(errorMessage))
 
-        val result = getLatestProductByKeywordUseCase(keyword = keyword)
+        val result = getFilteredProductsUseCase(postcode = samplePostcode)
 
-        assertNull(result)
+        assertTrue(result.isFailure)
+        val exception = result.exceptionOrNull()
+        assertTrue(exception is RuntimeException)
+        assertEquals(errorMessage, exception!!.message)
     }
 
     @Test
-    fun `GetLatestProductByKeywordUseCase should return null when no products match criteria`() = runTest {
+    fun `invoke should return empty list when no products match criteria`() = runTest {
         val productSummaries = listOf(
             ProductSummary(
                 code = "1",
@@ -139,10 +121,12 @@ class GetLatestProductByKeywordUseCaseTest {
                 brand = "OTHER_BRAND",
             ),
         )
+
         fakeRestApiRepository.setProductsResponse = Result.success(productSummaries)
 
-        val result = getLatestProductByKeywordUseCase(keyword = keyword)
+        val result = getFilteredProductsUseCase(postcode = samplePostcode)
 
-        assertNull(result)
+        assertTrue(result.isSuccess)
+        assertEquals(emptyList<ProductSummary>(), result.getOrNull())
     }
 }
