+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/product/GetStandardUnitRateUseCaseTest.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.usecase
+package com.rwmobi.kunigami.domain.usecase.product
 
 import com.rwmobi.kunigami.domain.repository.FakeOctopusApiRepository
 import com.rwmobi.kunigami.domain.repository.FakeUserPreferencesRepository
